feat(initial): first commit
This is the first commit to break out my work on improving
powerwall_monitor into it's own repository. This is being done so we can
simplify the process for users of powerwall_monitor to have it so they
only need to run the docker-compose file to bring up the whole network,
instead of also needing to build the containers.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>